requirements.txt
python-telegram-bot==20.3
SQLAlchemy==1.4.23
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
telethon==1.28.5
cryptg
pillow
aiohttp
